byOnTime/byBreakTime修改为uint6格式，led demo byOnTime/byBreakTime注释完善

+++ components/csi/include/drv/led.h
@@ -56,8 +56,8 @@ typedef struct csi_led_config {
 	uint8_t 	byClk;			//clk configure
 	uint8_t 	byBrt;			//brightness configure
 	uint16_t 	hwComMask;		//COM enable
-	uint8_t 	byOnTime;		//scanning timing: COM on cycles 
-	uint8_t 	byBreakTime;	//scanning timing: cycles between COMs
+	uint16_t 	byOnTime;		//scanning timing: COM on cycles 
+	uint16_t 	byBreakTime;	//scanning timing: cycles between COMs
 	uint8_t 	byInt;			//Interrupt Source
 }csi_led_config_t;
